qdrant fixes

+++ server/Hencoder/Services/QdrantServices/QdrantProxy.cs
@@ -17,28 +17,38 @@ namespace Hencoder.Services.QdrantServices
         public async Task<ulong> GetCollectionPointsCount()
         {
             using var client = new QdrantClient("localhost", 6334);
-            return (await client.GetCollectionInfoAsync(_collectionName)).PointsCount;
+            if (await client.CollectionExistsAsync(_collectionName))
+            {
+                return (await client.GetCollectionInfoAsync(_collectionName)).PointsCount;
+            }
+            return 0;
         }
 
         public async Task CreateCollection(Distance distance, ulong size)
         {
             using var client = new QdrantClient("localhost", 6334);
-            await client.CreateCollectionAsync(_collectionName,
-                new VectorParams
-                {
-                    Distance = distance,
-                    Size = size,
-                });
+            if (await client.CollectionExistsAsync(_collectionName) == false)
+            {
+                await client.CreateCollectionAsync(_collectionName,
+                    new VectorParams
+                    {
+                        Distance = distance,
+                        Size = size,
+                    });
+            }
         }
 
         public async Task Upsert(IEnumerable<QPoint> points)
         {
             using var client = new QdrantClient("localhost", 6334);
-            await client.UpsertAsync(_collectionName, points.Select(p => new PointStruct
+            if (await client.CollectionExistsAsync(_collectionName))
             {
-                Id = p.Id,
-                Vectors = p.Vector.ToArray(),
-            }).ToList());
+                await client.UpsertAsync(_collectionName, points.Select(p => new PointStruct
+                {
+                    Id = p.Id,
+                    Vectors = p.Vector.ToArray(),
+                }).ToList());
+            }
         }
 
         public async Task<IEnumerable<long>> Recommend(ulong N, float scoreThresholdValue, IEnumerable<long> positivePoints, IEnumerable<long> negativePoints, IEnumerable<long> notAllowedPoints = null)
@@ -65,18 +75,21 @@ namespace Hencoder.Services.QdrantServices
             {
                 query.Negative.Add(new PointId { Num = (ulong)p });
             }
-            using var client = new QdrantClient("localhost", 6334);
-            var points = await client.QueryAsync(
-                collectionName: _collectionName,
-                query: query,
-                filter: excludedPointsFilter,
-                limit: N,
-                scoreThreshold: scoreThresholdValue
-            );
             var set = new HashSet<long>();
-            foreach (var point in points)
+            using var client = new QdrantClient("localhost", 6334);
+            if (await client.CollectionExistsAsync(_collectionName))
             {
-                set.Add((long)point.Id.Num);
+                var points = await client.QueryAsync(
+                    collectionName: _collectionName,
+                    query: query,
+                    filter: excludedPointsFilter,
+                    limit: N,
+                    scoreThreshold: scoreThresholdValue
+                );                
+                foreach (var point in points)
+                {
+                    set.Add((long)point.Id.Num);
+                }
             }
             return set;
         }
@@ -84,7 +97,11 @@ namespace Hencoder.Services.QdrantServices
         public async Task DeleteCollection()
         {
             using var client = new QdrantClient("localhost", 6334);
-            await client.DeleteCollectionAsync(_collectionName);
+            if (await client.CollectionExistsAsync(_collectionName))
+            {
+                await client.DeleteCollectionAsync(_collectionName);
+            }
+
         }
 
         public void Dispose()
